requirements.txt
Mesa>=3.0.0
Mesa-LLM>=0.1.0
pydantic>=2.0.0
litellm>=1.0.0
rich>=13.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
PyYAML>=6.0.0
